예수금 Float 타입으로 변경

+++ Virtual Investment/Virtual Investment/DataBase/AmountData.swift
@@ -9,7 +9,7 @@ import Foundation
 
 class AmountData {
   static var shared = AmountData()
-  var deposit: Int = 0
+  var deposit: Float = 0
   var evaluatedPrice: Float = 0
   var investmentAccount: Float = 0
   var investededCoins: [Coin] = []
